backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
faiss-cpu==1.13.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
tqdm==4.66.1
openai==1.40.0
